GNU find (i.e findutils) command for at least version 4.4.2 spills out last modified timestamp in option -printf "%T@\n" with fractional parts. GNU find version 4.2.27 output are integer without fractional part. For that reason tryParseLastChangedMillis() parses the timestamp as a double value.
SVN: 23827

+++ datamover/source/java/ch/systemsx/cisd/datamover/filesystem/store/FileStoreRemote.java
@@ -588,7 +588,7 @@ public class FileStoreRemote extends AbstractFileStore
     {
         try
         {
-            return Long.parseLong(numberStr) * 1000;
+            return (long) (Double.parseDouble(numberStr) * 1000);
         } catch (final NumberFormatException e)
         {
             return null;
